+++ source/Engine/Materials/LightMaterial.cpp
@@ -38,7 +38,7 @@ void LightMaterial::UploadDerivedVariables()
 	glUniform1i(glGetUniformLocation(m_Shader->GetProgram(), "texGBufferB"), 1);
 	glUniform1i(glGetUniformLocation(m_Shader->GetProgram(), "texGBufferC"), 2);
 	auto gbufferTex = RenderPipeline::GetInstance()->GetGBuffer()->GetTextures();
-	for (size_t i = 0; i < gbufferTex.size(); i++)
+	for (uint32 i = 0; i < (uint32)gbufferTex.size(); i++)
 	{
 		STATE->LazyBindTexture(i, GL_TEXTURE_2D, gbufferTex[i]->GetHandle());
 	}
